requirements.txt
django
djangorestframework==3.14.0
Pillow==10.0.1